mainly formating

+++ src/convert.h
@@ -38,28 +38,28 @@ public:
 
     ~Convert();
 
-    static QString num( long int num );
-    static QString bytes( const QString& x );
-    static QString bytes( long int x );
-    static QString bytes( double x, int precision = 2 );
-    static QString bytes( qulonglong x );
-    static QString bytes( const QString& x, const QString& y );
-    static QString bytes( qulonglong x, int precision );
-    static QString bytesLong( const QString& x );
-    static QString bytesExtra( const QString& x );
-    static QString power( const QString& x );
-    static float powerValue( const QString& x );
-    static QString power( float power );
-    static QString time( long int seconds );
+    static QString num(long int num);
+    static QString bytes(const QString& x);
+    static QString bytes(long int x);
+    static QString bytes(double x, int precision = 2);
+    static QString bytes(qulonglong x);
+    static QString bytes(const QString& x, const QString& y);
+    static QString bytes(qulonglong x, int precision);
+    static QString bytesLong(const QString& x);
+    static QString bytesExtra(const QString& x);
+    static QString power(const QString& x);
+    static float powerValue(const QString& x);
+    static QString power(float power);
+    static QString time(long int seconds);
     static QString getFileErrorString(QFile::FileError error);
 
-    static const int ONE_DAY = 86400;
-    static const int ONE_HOUR = 3600;
-    static const int ONE_MINUTE = 60;
+    static const long ONE_DAY = 86400;
+    static const long ONE_HOUR = 3600;
+    static const long ONE_MINUTE = 60;
 
-    static const int ONE_GIG = 1073741824;
-    static const int ONE_MEG = 1048576;
-    static const int ONE_KILO = 1024;
+    static const long ONE_GIG = 1073741824;
+    static const long ONE_MEG = 1048576;
+    static const long ONE_KILO = 1024;
 
     template<class X1> static const X1& min(const X1& a, const X1& b) {
         if(a < b) {
